fix(csa-config): add prefix matching and "Did you mean?" for tier selectors
resolve_tier_selector() now tries unambiguous prefix matching after exact
name and alias lookup both fail. If the selector uniquely prefixes exactly
one tier name (e.g. "tier-1" → "tier-1-quick"), it resolves automatically.
Ambiguous prefixes still return None.

Also adds suggest_tier() for error messages: when tier resolution fails,
the error now includes "Did you mean 'tier-1-quick'?" when a prefix or
substring match is found. Updated all 3 call sites (run, review, debate).

Test file config_tests_tier.rs split into config_tests_tier_selector.rs
to stay under the 800-line monolith limit.

Closes #458
