+++ crates/cli-sub-agent/src/run_helpers.rs
@@ -75,9 +75,13 @@ pub(crate) fn resolve_tool_and_model(
             } else {
                 let available: Vec<&str> = cfg.tiers.keys().map(|k| k.as_str()).collect();
                 let alias_hint = cfg.format_tier_aliases();
+                let suggest_hint = cfg
+                    .suggest_tier(tier_name)
+                    .map(|s| format!("\nDid you mean '{s}'?"))
+                    .unwrap_or_default();
                 anyhow::bail!(
                     "Tier selector '{}' not found.\n\
-                     Available tiers: [{}]{alias_hint}",
+                     Available tiers: [{}]{alias_hint}{suggest_hint}",
                     tier_name,
                     available.join(", ")
                 );
