+++ crates/csa-config/src/config_tests_tier.rs
@@ -538,202 +538,3 @@ fn filtered_returns_none_when_all_restricted_and_needs_edit() {
     let result = config.resolve_tier_tool_filtered("default", true);
     assert!(result.is_none());
 }
-
-// --- resolve_tier_selector tests ---
-
-#[test]
-fn test_resolve_tier_selector_direct_tier() {
-    let mut tiers = HashMap::new();
-    tiers.insert(
-        "tier1".to_string(),
-        TierConfig {
-            description: "test".to_string(),
-            models: vec!["gemini-cli/google/default/xhigh".to_string()],
-            strategy: TierStrategy::default(),
-            token_budget: None,
-            max_turns: None,
-        },
-    );
-
-    let config = ProjectConfig {
-        schema_version: CURRENT_SCHEMA_VERSION,
-        project: ProjectMeta::default(),
-        resources: ResourcesConfig::default(),
-        acp: Default::default(),
-        tools: HashMap::new(),
-        review: None,
-        debate: None,
-        tiers,
-        tier_mapping: HashMap::new(),
-        aliases: HashMap::new(),
-        tool_aliases: HashMap::new(),
-        preferences: None,
-        session: Default::default(),
-        memory: Default::default(),
-        hooks: Default::default(),
-        execution: Default::default(),
-        vcs: Default::default(),
-        filesystem_sandbox: Default::default(),
-    };
-
-    assert_eq!(
-        config.resolve_tier_selector("tier1"),
-        Some("tier1".to_string())
-    );
-}
-
-#[test]
-fn test_resolve_tier_selector_alias() {
-    let mut tiers = HashMap::new();
-    tiers.insert(
-        "tier-2-standard".to_string(),
-        TierConfig {
-            description: "test".to_string(),
-            models: vec!["gemini-cli/google/default/xhigh".to_string()],
-            strategy: TierStrategy::default(),
-            token_budget: None,
-            max_turns: None,
-        },
-    );
-
-    let mut tier_mapping = HashMap::new();
-    tier_mapping.insert("default".to_string(), "tier-2-standard".to_string());
-
-    let config = ProjectConfig {
-        schema_version: CURRENT_SCHEMA_VERSION,
-        project: ProjectMeta::default(),
-        resources: ResourcesConfig::default(),
-        acp: Default::default(),
-        tools: HashMap::new(),
-        review: None,
-        debate: None,
-        tiers,
-        tier_mapping,
-        aliases: HashMap::new(),
-        tool_aliases: HashMap::new(),
-        preferences: None,
-        session: Default::default(),
-        memory: Default::default(),
-        hooks: Default::default(),
-        execution: Default::default(),
-        vcs: Default::default(),
-        filesystem_sandbox: Default::default(),
-    };
-
-    assert_eq!(
-        config.resolve_tier_selector("default"),
-        Some("tier-2-standard".to_string())
-    );
-}
-
-#[test]
-fn test_resolve_tier_selector_direct_wins_on_collision() {
-    let mut tiers = HashMap::new();
-    // A tier named "default" exists directly
-    tiers.insert(
-        "default".to_string(),
-        TierConfig {
-            description: "direct tier".to_string(),
-            models: vec!["gemini-cli/google/default/xhigh".to_string()],
-            strategy: TierStrategy::default(),
-            token_budget: None,
-            max_turns: None,
-        },
-    );
-    tiers.insert(
-        "tier-2-standard".to_string(),
-        TierConfig {
-            description: "standard".to_string(),
-            models: vec!["codex/openai/gpt-5/xhigh".to_string()],
-            strategy: TierStrategy::default(),
-            token_budget: None,
-            max_turns: None,
-        },
-    );
-
-    let mut tier_mapping = HashMap::new();
-    // tier_mapping also maps "default" to a different tier
-    tier_mapping.insert("default".to_string(), "tier-2-standard".to_string());
-
-    let config = ProjectConfig {
-        schema_version: CURRENT_SCHEMA_VERSION,
-        project: ProjectMeta::default(),
-        resources: ResourcesConfig::default(),
-        acp: Default::default(),
-        tools: HashMap::new(),
-        review: None,
-        debate: None,
-        tiers,
-        tier_mapping,
-        aliases: HashMap::new(),
-        tool_aliases: HashMap::new(),
-        preferences: None,
-        session: Default::default(),
-        memory: Default::default(),
-        hooks: Default::default(),
-        execution: Default::default(),
-        vcs: Default::default(),
-        filesystem_sandbox: Default::default(),
-    };
-
-    // Direct tier name wins over mapping alias
-    assert_eq!(
-        config.resolve_tier_selector("default"),
-        Some("default".to_string())
-    );
-}
-
-#[test]
-fn test_resolve_tier_selector_nonexistent() {
-    let config = ProjectConfig {
-        schema_version: CURRENT_SCHEMA_VERSION,
-        project: ProjectMeta::default(),
-        resources: ResourcesConfig::default(),
-        acp: Default::default(),
-        tools: HashMap::new(),
-        review: None,
-        debate: None,
-        tiers: HashMap::new(),
-        tier_mapping: HashMap::new(),
-        aliases: HashMap::new(),
-        tool_aliases: HashMap::new(),
-        preferences: None,
-        session: Default::default(),
-        memory: Default::default(),
-        hooks: Default::default(),
-        execution: Default::default(),
-        vcs: Default::default(),
-        filesystem_sandbox: Default::default(),
-    };
-
-    assert_eq!(config.resolve_tier_selector("unknown"), None);
-}
-
-#[test]
-fn test_resolve_tier_selector_alias_to_nonexistent_tier() {
-    let mut tier_mapping = HashMap::new();
-    tier_mapping.insert("broken".to_string(), "nonexistent-tier".to_string());
-
-    let config = ProjectConfig {
-        schema_version: CURRENT_SCHEMA_VERSION,
-        project: ProjectMeta::default(),
-        resources: ResourcesConfig::default(),
-        acp: Default::default(),
-        tools: HashMap::new(),
-        review: None,
-        debate: None,
-        tiers: HashMap::new(),
-        tier_mapping,
-        aliases: HashMap::new(),
-        tool_aliases: HashMap::new(),
-        preferences: None,
-        session: Default::default(),
-        memory: Default::default(),
-        hooks: Default::default(),
-        execution: Default::default(),
-        vcs: Default::default(),
-        filesystem_sandbox: Default::default(),
-    };
-
-    assert_eq!(config.resolve_tier_selector("broken"), None);
-}
